xenpm: Add set-cpufreq-hwp subcommand
set-cpufreq-hwp allows setting the Hardware P-State (HWP) parameters.

It can be run on all or just a snigle cpu.  There are preset of
balance, powersave & performance.  Those can be further tweaked by
param:val arguments as explained in the usage description.

parameter names are just checked to the first 3 characters to shorten
typing.

Some options are hardware dependent, and ranges can be found in
get-cpufreq-para.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>